Add Debug Maze Generation to Prims
